third_party/WebKit/LayoutTests/platform/win/http/tests/security/sandbox-iframe-blocks-top-navigation-expected.txt

CONSOLE ERROR: line 3: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/security/sandbox-iframe-blocks-top-navigation.html' from frame with URL 'about:srcdoc'. The frame attempting navigation of the top-level window is sandboxed, but the flag of 'allow-top-navigation' or 'allow-top-navigation-by-user-activation' is not set.

This is a testharness.js-based test.
PASS Sandboxed frames should throw when navigating the top-level window. 
Harness: the test ran to completion.
